tests/test_backup_restore.py
import pytest
import os
import yaml
from pathlib import Path


def test_backup_config():
    # Test backup configuration
    backup_dir = Path("backup")
    assert backup_dir.exists()

    # Test backup script
    backup_script = backup_dir / "backup.sh"
    assert backup_script.exists()
    assert os.access(backup_script, os.X_OK)

    with open(backup_script, "r") as f:
        content = f.read()

        # Verify backup commands
        assert "pg_dump" in content
        assert "tar -czf" in content
        assert "aws s3 cp" in content

        # Verify backup paths
        assert "backup/database" in content
        assert "backup/files" in content
        assert "backup/logs" in content


def test_restore_config():
    # Test restore configuration
    backup_dir = Path("backup")

    # Test restore script
    restore_script = backup_dir / "restore.sh"
    assert restore_script.exists()
    assert os.access(restore_script, os.X_OK)

    with open(restore_script, "r") as f:
        content = f.read()

        # Verify restore commands
        assert "pg_restore" in content
        assert "tar -xzf" in content
        assert "aws s3 cp" in content

        # Verify restore paths
        assert "backup/database" in content
        assert "backup/files" in content
        assert "backup/logs" in content


def test_backup_schedule():
    # Test backup schedule
    backup_dir = Path("backup")

    # Test cron configuration
    cron_path = backup_dir / "cron"
    assert cron_path.exists()

    with open(cron_path, "r") as f:
        content = f.read()

        # Verify cron schedule
        assert "0 0 * * *" in content  # Daily backup
        assert "0 0 * * 0" in content  # Weekly backup
        assert "0 0 1 * *" in content  # Monthly backup


def test_backup_retention():
    # Test backup retention
    backup_dir = Path("backup")

    # Test retention policy
    retention_path = backup_dir / "retention.yaml"
    assert retention_path.exists()

    with open(retention_path, "r") as f:
        retention = yaml.safe_load(f)

        # Verify retention policy
        assert "daily" in retention
        assert "weekly" in retention
        assert "monthly" in retention

        # Verify retention periods
        assert retention["daily"] > 0
        assert retention["weekly"] > 0
        assert retention["monthly"] > 0


def test_backup_verification():
    # Test backup verification
    backup_dir = Path("backup")

    # Test verification script
    verify_script = backup_dir / "verify.sh"
    assert verify_script.exists()
    assert os.access(verify_script, os.X_OK)

    with open(verify_script, "r") as f:
        content = f.read()

        # Verify verification commands
        assert "pg_restore" in content
        assert "md5sum" in content
        assert "diff" in content


def test_backup_monitoring():
    # Test backup monitoring
    backup_dir = Path("backup")

    # Test monitoring script
    monitor_script = backup_dir / "monitor.sh"
    assert monitor_script.exists()
    assert os.access(monitor_script, os.X_OK)

    with open(monitor_script, "r") as f:
        content = f.read()

        # Verify monitoring commands
        assert "find" in content
        assert "du" in content
        assert "curl" in content


def test_backup_notification():
    # Test backup notification
    backup_dir = Path("backup")

    # Test notification script
    notify_script = backup_dir / "notify.sh"
    assert notify_script.exists()
    assert os.access(notify_script, os.X_OK)

    with open(notify_script, "r") as f:
        content = f.read()

        # Verify notification commands
        assert "curl" in content
        assert "mail" in content
        assert "slack" in content


def test_backup_encryption():
    # Test backup encryption
    backup_dir = Path("backup")

    # Test encryption script
    encrypt_script = backup_dir / "encrypt.sh"
    assert encrypt_script.exists()
    assert os.access(encrypt_script, os.X_OK)

    with open(encrypt_script, "r") as f:
        content = f.read()

        # Verify encryption commands
        assert "gpg" in content
        assert "openssl" in content
        assert "chmod 600" in content
